requirements.txt
cassandra_driver==3.29.2
